% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.gene.choose.k.R
\name{nmf.gene.choose.k}
\alias{nmf.gene.choose.k}
\title{A Function to find responsible genes for variation within a variable}
\usage{
nmf.gene.choose.k(k.range = seq(2, ncol(gene_expressions)/10, 2),
  gene_expressions, zero.threshold = 0.1, log.scaled = T,
  loss.type = "mse", b = c(0, 0, 0), na = FALSE, sample = FALSE,
  number.genes, a = c(0, 0, 0))
}
\arguments{
\item{k.range}{Suggest a vector of k values to explore, with a default}

\item{zero.threshold}{what proportion of zeros should a gene have before being left out of the analysis}

\item{number.genes}{The number of likely genes to find}

\item{variable}{The discrete variable or disease to be investigated}

\item{gene.names}{A character/string vector containing the list of gene names}

\item{gene.expressions}{The initial gene expression matrix, with genes labelled by row names}
}
\description{
This function allows you to find the likely genes responsible for affecting a variable, typically the presence or absence of a disease
}
\examples{
nmf.responsible.genes()
}
\keyword{cats}
