% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.heatmap.R
\name{nmf.heatmap}
\alias{nmf.heatmap}
\title{A Function to find responsible genes for variation within a variable}
\usage{
nmf.heatmap(H.matrix, variable, step = 1,
  variable.name = "unspecified variable")
}
\arguments{
\item{H.matrix}{The H.matrix of an NMF factorisation of a gene expression dataset}

\item{variable}{A vector containing values for the discrete variable or disease to be investigated}

\item{step}{The skip value to determine how many patients/subjects to display}

\item{variable.name}{The name of the variable being investigated}
}
\description{
This function allows you to find the likely genes responsible for affecting a variable, typically the presence or absence of a disease
}
\examples{
nmf.responsible.genes()
}
\keyword{cats}
