% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.responsible.genes.R
\name{nmf.responsible.genes}
\alias{nmf.responsible.genes}
\title{A Function to find responsible genes for variation within a variable}
\usage{
nmf.responsible.genes(H.matrix, W.matrix, variable, number.genes, gene.names,
  gene.expressions, control.age = T, age = NA)
}
\arguments{
\item{H.matrix}{The H.matrix of an NMF factorisation of a gene expression dataset}

\item{W.matrix}{The W.matrix of an NMF factorisation of a gene expression dataset}

\item{variable}{The discrete variable or disease to be investigated}

\item{number.genes}{The number of likely genes to find}

\item{gene.names}{A character/string vector containing the list of gene names}

\item{gene.expressions}{The initial gene expression matrix, with genes labelled by row names}
}
\description{
This function allows you to find the likely genes responsible for affecting a variable, typically the presence or absence of a disease
}
\examples{
nmf.responsible.genes()
}
\keyword{cats}
